requirements.txt
tensorflow==2.7.0
matplotlib==3.4.3
pillow==8.2.0